+++ src/common/storage/Cargo.toml
@@ -20,6 +20,7 @@ common-meta-app = { path = "../../meta/app" }
 anyhow = { workspace = true }
 async-trait = "0.1"
 bytes = "1"
+chrono = { workspace = true }
 futures = "0.3"
 opendal = { workspace = true }
 regex = "1.6.0"
