--- src/query/catalog/src/plan/stage_file_info.rs
@@ -1,34 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use chrono::DateTime;
-use chrono::Utc;
-use common_meta_app::principal::UserIdentity;
-
-#[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub enum StageFileStatus {
-    NeedCopy,
-    AlreadyCopied,
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct StageFileInfo {
-    pub path: String,
-    pub size: u64,
-    pub md5: Option<String>,
-    pub last_modified: DateTime<Utc>,
-    pub etag: Option<String>,
-    pub status: StageFileStatus,
-    pub creator: Option<UserIdentity>,
-}
