+++ src/query/pipeline/sources/Cargo.toml
@@ -24,6 +24,7 @@ common-io = { path = "../../../common/io" }
 common-meta-app = { path = "../../../meta/app" }
 common-pipeline-core = { path = "../core" }
 common-settings = { path = "../../settings" }
+common-storage = { path = "../../../common/storage" }
 
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
 bstr = "1.0.1"
