+++ src/query/pipeline/sources/src/input_formats/impls/input_format_parquet.rs
@@ -35,7 +35,6 @@ use common_arrow::parquet::metadata::RowGroupMetaData;
 use common_arrow::parquet::read::read_metadata;
 use common_arrow::read_columns_async;
 use common_base::runtime::execute_futures_in_parallel;
-use common_catalog::plan::StageFileInfo;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::DataBlock;
@@ -45,6 +44,7 @@ use common_expression::TableSchemaRef;
 use common_meta_app::principal::StageInfo;
 use common_pipeline_core::Pipeline;
 use common_settings::Settings;
+use common_storage::StageFileInfo;
 use futures::AsyncRead;
 use futures::AsyncSeek;
 use opendal::Operator;
