+++ src/query/service/src/interpreters/interpreter_user_stage_remove.rs
@@ -17,8 +17,8 @@ use std::sync::Arc;
 use common_catalog::table_context::TableContext;
 use common_exception::Result;
 use common_sql::plans::RemoveStagePlan;
+use common_storage::StageFilesInfo;
 use common_storages_fuse::io::Files;
-use common_storages_stage::list_file;
 use common_storages_stage::StageTable;
 use tracing::error;
 
@@ -48,11 +48,25 @@ impl Interpreter for RemoveUserStageInterpreter {
     async fn execute2(&self) -> Result<PipelineBuildResult> {
         let plan = self.plan.clone();
         let op = StageTable::get_op(&self.plan.stage)?;
-        let files = list_file(&op, &plan.path, &plan.pattern).await?;
+        let pattern = if plan.pattern.is_empty() {
+            None
+        } else {
+            Some(plan.pattern.clone())
+        };
+        let files_info = StageFilesInfo {
+            path: plan.path.clone(),
+            files: None,
+            pattern,
+        };
+        let files: Vec<String> = files_info
+            .list(&op, false)
+            .await?
+            .into_iter()
+            .map(|file_with_meta| file_with_meta.path)
+            .collect::<Vec<_>>();
 
         let table_ctx: Arc<dyn TableContext> = self.ctx.clone();
         let file_op = Files::create(table_ctx, op);
-        let files = files.iter().map(|v| v.path.clone()).collect::<Vec<_>>();
         if let Err(e) = file_op.remove_file_in_batch(&files).await {
             error!("Failed to delete file: {:?}, error: {}", files, e);
         }
